fix: SRE `Available Qty to Reserve` for Group Warehouse
